+++ data.js
@@ -1,0 +1,48 @@
+const countries = [
+  'Argentina',
+  'Andorra',
+  'Austria',
+  'Bahamas',
+  'Belgium',
+  'Bolivia',
+  'Brazil',
+  'Canada',
+  'Chile',
+  'Colombia',
+  'Costa Rica',
+  'Croatia',
+  'Czechia',
+  'Denmark',
+  'Ecuador',
+  'England',
+  'Estonia',
+  'Finland',
+  'France',
+  'Germany',
+  'Hungary',
+  'Ireland',
+  'Italy',
+  'Latvia',
+  'Lithuania',
+  'Luxembourg',
+  'Mexico',
+  'Monaco',
+  'Morocco',
+  'Netherlands',
+  'Norway',
+  'Peru',
+  'Poland',
+  'Portugal',
+  'Qatar',
+  'Slovakia',
+  'Slovenia',
+  'Spain',
+  'Switzerland',
+  'Turkey',
+  'United Arab Emirates',
+  'United States',
+  'Uruguay',
+  'Venezuela'
+];
+
+export default { countries };
