+++ routes/home.js
@@ -1,0 +1,8 @@
+import express from 'express';
+const router = express.Router();
+
+router.get("/", (req, res) => {
+  res.render('index.ejs');
+});
+
+export default router;
