+++ views/destinations.ejs
@@ -4,7 +4,7 @@
   <h2 class="pb-2 text-center display-4">Destinations</h2>
   <div class="container">
     <div class="row row-cols-1 row-cols-lg-3 align-items-stretch g-4 py-5">
-      <% countries.forEach(country => { %>
+      <% countries.countries.forEach(country => { %>
         <div class="col">
           <div class="card card-cover h-100 overflow-hidden text-bg-dark rounded-4 shadow-lg">
             <div class="d-flex flex-column h-100 p-5 pb-3 text-white text-shadow-1">
